Upgrade to sidekiq-unique-jobs version 7
This is a major overhaul of the sidekiq-unique-jobs plugin. In addition to API changes, it introduces a "reaper" feature that automatically searches for and removes orphan locks (every 10 minutes by default). This means we can remove our code that interacts with the locks.

We'll need to keep careful track of this in production to make sure it's working as expected.