Smoother handling of missing default campaign
Instead of raising an error or trying to automatically create the default campaign if it doesn't exist, this lets the home and explore pages just work without any default campaign.

If the default campaign doesn't exist, the explore page will redirect to the home page (and the home page won't include the stats from the default campaign).