Automate sending term recap emails when each course ends
This automates sending the term recap emails for each course in Wiki Education's Wikipedia Student Program after it ends, so that we don't need to periodically use the manual mass_emails interface to kick off a batch of these emails.

It will only email courses within a few days of the end of the course, and should be compatible to work alongside the mass_email feature (so we can use that to email courses from this term that are already out of the sending window, before we let automation take over).